refactor(loco-base): share _obs_noise and BaseNoiseConfig
