+++ src/unilab/envs/locomotion/common/base.py
@@ -26,8 +26,19 @@ class ControlConfigBase:
 
 
 @dataclass
+class BaseNoiseConfig:
+    level: float = 0.0
+    scale_joint_angle: float = 0.03
+    scale_joint_vel: float = 0.5
+    scale_gyro: float = 0.2
+    scale_gravity: float = 0.05
+    scale_linvel: float = 0.1
+
+
+@dataclass
 class LocomotionBaseCfg(EnvCfg):
     control_config: ControlConfigBase = field(default_factory=ControlConfigBase)
+    noise_config: BaseNoiseConfig = field(default_factory=BaseNoiseConfig)
     sensor: Sensor = field(default_factory=Sensor)
     sim_dt: float = 0.01
     ctrl_dt: float = 0.02
@@ -81,6 +92,14 @@ class LocomotionBaseEnv(NpEnv):
             exec_actions * self._cfg.control_config.action_scale + self.default_angles
         )
         return ctrl
+
+    def _obs_noise(self, data: np.ndarray, scale: float) -> np.ndarray:
+        """Apply per-step uniform observation noise scaled by ``noise_config.level``."""
+        level = float(self._cfg.noise_config.level)
+        if level <= 0.0:
+            return data
+        noise = np.random.uniform(-1.0, 1.0, data.shape).astype(data.dtype) * level * scale
+        return data + noise
 
     def get_local_linvel(self) -> np.ndarray:
         local_linvel: np.ndarray = self._backend.get_sensor_data(self._cfg.sensor.local_linvel)
