+++ src/unilab/envs/locomotion/g1/base.py
@@ -5,6 +5,7 @@ from dataclasses import dataclass, field
 import numpy as np
 
 from unilab.envs.locomotion.common.base import (
+    BaseNoiseConfig,
     ControlConfigBase,
     LocomotionBaseCfg,
     LocomotionBaseEnv,
@@ -15,13 +16,10 @@ from unilab.envs.locomotion.common.base import (
 
 
 @dataclass
-class NoiseConfig:
-    level: float = 0.0
+class NoiseConfig(BaseNoiseConfig):
     scale_joint_angle: float = 0.02
     scale_joint_vel: float = 0.3
     scale_gyro: float = 0.1
-    scale_gravity: float = 0.05
-    scale_linvel: float = 0.1
 
 
 @dataclass
@@ -45,7 +43,7 @@ class Asset:
 
 @dataclass
 class G1BaseCfg(LocomotionBaseCfg):
-    noise_config: NoiseConfig = field(default_factory=NoiseConfig)
+    noise_config: NoiseConfig = field(default_factory=NoiseConfig)  # type: ignore[assignment]
     control_config: ControlConfig = field(default_factory=ControlConfig)  # type: ignore[assignment]
     sensor: Sensor = field(default_factory=Sensor)
     asset: Asset = field(default_factory=Asset)
@@ -59,16 +57,5 @@ class G1BaseEnv(LocomotionBaseEnv):
     _use_global_dtype = False
 
     def _obs_noise(self, data: np.ndarray, scale: float) -> np.ndarray:
-        """Apply per-step uniform observation noise scaled by ``noise_config.level``."""
-        noise_cfg = self._cfg.noise_config
-        if noise_cfg.level > 0.0:
-            return np.asarray(
-                data
-                + (
-                    np.random.uniform(-1.0, 1.0, data.shape).astype(data.dtype)
-                    * noise_cfg.level
-                    * scale
-                ),
-                dtype=data.dtype,
-            )
-        return np.asarray(data)
+        """Same as base, but coerces back to ``data.dtype`` (G1 runs in float32)."""
+        return np.asarray(super()._obs_noise(data, scale), dtype=data.dtype)
