+++ src/unilab/envs/locomotion/go1/base.py
@@ -2,9 +2,8 @@ from __future__ import annotations
 
 from dataclasses import dataclass, field
 
-import numpy as np
-
 from unilab.envs.locomotion.common.base import (
+    BaseNoiseConfig,
     ControlConfigBase,
     LocomotionBaseCfg,
     LocomotionBaseEnv,
@@ -12,13 +11,8 @@ from unilab.envs.locomotion.common.base import (
 
 
 @dataclass
-class NoiseConfig:
-    level: float = 0.0
-    scale_joint_angle: float = 0.03
-    scale_joint_vel: float = 0.5
-    scale_gyro: float = 0.2
-    scale_gravity: float = 0.05
-    scale_linvel: float = 0.1
+class NoiseConfig(BaseNoiseConfig):
+    pass
 
 
 @dataclass
@@ -36,7 +30,7 @@ class Asset:
 
 @dataclass
 class Go1BaseCfg(LocomotionBaseCfg):
-    noise_config: NoiseConfig = field(default_factory=NoiseConfig)
+    noise_config: NoiseConfig = field(default_factory=NoiseConfig)  # type: ignore[assignment]
     control_config: ControlConfig = field(default_factory=ControlConfig)  # type: ignore[assignment]
     asset: Asset = field(default_factory=Asset)
     sim_dt: float = 0.01
@@ -45,14 +39,3 @@ class Go1BaseCfg(LocomotionBaseCfg):
 
 class Go1BaseEnv(LocomotionBaseEnv):
     _cfg: Go1BaseCfg
-
-    def _obs_noise(self, data: np.ndarray, scale: float) -> np.ndarray:
-        """Apply per-step uniform observation noise scaled by ``noise_config.level``."""
-        noise_cfg = self._cfg.noise_config
-        if noise_cfg.level > 0.0:
-            return data + (
-                np.random.uniform(-1.0, 1.0, data.shape).astype(data.dtype)
-                * noise_cfg.level
-                * scale
-            )
-        return data
